# api/openapi.yaml
openapi: 3.0.3
info:
  title: boxincloud API
  version: 0.1.0-dev
  description: |
    Bibliothèque de BD, comics et mangas sur stockage objet.

    Contrat de l'API boxincloud.

    **Ce fichier est la source de vérité.** Le serveur Go, le client TypeScript
    du web et le client Dart de l'application Flutter en sont générés
    (`make generate`). Toute évolution de l'API commence ici — jamais par le
    code, sinon les clients divergent.

    Un test de contrat (`internal/httpapi/contract_test.go`) valide chaque
    réponse du serveur réel contre ce schéma : une divergence entre le contrat
    et l'implémentation fait échouer la CI.

    ## Erreurs

    Les erreurs suivent la [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457)
    (`application/problem+json`). Le champ `type` est stable : les clients
    peuvent s'appuyer dessus pour brancher leur comportement. Deux valeurs
    méritent un traitement particulier :

    - `.../token-expired` — le jeton d'accès a expiré, rafraîchir puis rejouer ;
    - `.../session-revoked` — la session est révoquée, redemander une connexion
      complète (une réutilisation de refresh token a été détectée).

    ## Pagination

    Pagination par curseur. Le client renvoie `nextCursor` tel quel dans le
    paramètre `cursor` ; il ne doit jamais le construire ni l'interpréter.
    L'absence de `nextCursor` signifie « dernière page ».

    ## Versionnement

    Le préfixe `/api/v1` change uniquement en cas de rupture. Les ajouts
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/setup:
    post:
      operationId: setup
      summary: Créer le premier administrateur
      description: |
        Ouvert sans authentification tant qu'aucun compte n'existe — c'est le
        seul moyen d'amorcer une instance neuve. Répond 403 dès qu'un compte a
        été créé.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                  description: Lettres, chiffres, tiret, point et souligné.
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
                  format: password
      responses:
        "201":
          description: Administrateur créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /auth/login:
    post:
      operationId: login
      summary: Ouvrir une session
      description: |
        Retourne un jeton d'accès de courte durée et un refresh token.

        `deviceId` est facultatif au premier appel : le serveur en crée un et le
        retourne. Le client le conserve pour que ses reconnexions soient
        rattachées au même appareil.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
                deviceId:
                  type: string
                  format: uuid
                  description: Appareil existant. Omis, un nouvel appareil est créé.
                deviceName:
                  type: string
                  description: Nom lisible, affiché dans la liste des appareils.
                platform:
                  $ref: "#/components/schemas/DevicePlatform"
                appVersion: { type: string }
      responses:
        "200":
          description: Session ouverte
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tokens" }
        "401":
          description: |
            Identifiants invalides. La réponse est identique que le compte
            n'existe pas ou que le mot de passe soit erroné — distinguer les
            deux permettrait d'énumérer les comptes.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /auth/refresh:
    post:
      operationId: refreshTokens
      summary: Renouveler les jetons
      description: |
        Échange un refresh token contre un nouveau couple. Le jeton présenté est
        révoqué — rotation systématique.

        Si un jeton déjà tourné est présenté, c'est qu'il a été volé : toute la
        chaîne de rotation est révoquée et la réponse porte le type
        `.../session-revoked`. Le client doit alors redemander une connexion
        complète.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Nouveaux jetons
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tokens" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /auth/logout:
    post:
      operationId: logout
      summary: Fermer une session
      description: |
        Idempotent : un jeton inconnu, déjà révoqué ou expiré n'est pas une
        erreur. Un client hors ligne au moment de la déconnexion peut rejouer sa
        requête plus tard.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: Session fermée
  /me:
    get:
      operationId: getCurrentUser
      summary: Compte courant
      tags: [auth]
      responses:
        "200":
          description: Compte courant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/devices:
    get:
      operationId: listDevices
      summary: Appareils connectés
      tags: [auth]
      responses:
        "200":
          description: Liste des appareils
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: "#/components/schemas/Device" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/logout-all:
    post:
      operationId: logoutAllDevices
      summary: Révoquer toutes les sessions
      tags: [auth]
      responses:
        "200":
          description: Sessions révoquées
          content:
            application/json:
              schema:
                type: object
                required: [revokedSessions]
                properties:
                  revokedSessions: { type: integer, format: int64 }
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Catalogue ──────────────────────────────────────────────────────────────
  /home:
    get:
      operationId: getHome
      summary: Étagères d'accueil
      description: |
        Regroupe les étagères de la page d'accueil. « Reprendre la lecture » est
        servi séparément par `/continue-reading` : il relève de la progression,
        dont le catalogue ne dépend pas.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Étagères
          content:
            application/json:
              schema:
                type: object
                required: [recent, nextInSeries]
                properties:
                  recent:
                    type: array
                    items: { $ref: "#/components/schemas/Comic" }
                  nextInSeries:
                    description: Premier album non lu de chaque série entamée.
                    type: array
                    items: { $ref: "#/components/schemas/Comic" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search:
    get:
      operationId: search
      summary: Rechercher dans la bibliothèque
      description: |
        Recherche plein texte combinée à une similarité trigramme, insensible
        aux accents : « asterix » trouve « Astérix », « asterics » aussi.

        Une requête de moins de deux caractères retourne des listes vides.
      tags: [catalog]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: Termes recherchés.
        - $ref: "#/components/parameters/LibraryId"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Résultats
          content:
            application/json:
              schema:
                type: object
                required: [comics, series]
                properties:
                  comics:
                    type: array
                    items: { $ref: "#/components/schemas/Comic" }
                  series:
                    type: array
                    items: { $ref: "#/components/schemas/Series" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /libraries:
    post:
      operationId: createLibrary
      summary: Créer une bibliothèque sur un backend existant
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, backendId]
              properties:
                name: { type: string }
                backendId: { type: string, format: uuid }
                kind:
                  type: string
                  description: "Défaut : comics."
                rootPrefix:
                  type: string
                  description: |
                    Préfixe sous lequel la bibliothèque vit dans le backend.
                    Vide pour la racine.
      responses:
        "201":
          description: Bibliothèque créée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    get:
      operationId: listLibraries
      summary: Bibliothèques visibles
      description: |
        Seules les bibliothèques auxquelles le compte a accès sont retournées.
        Une bibliothèque sans restriction est visible de tous.
      tags: [catalog]
      responses:
        "200":
          description: Bibliothèques
          content:
            application/json:
              schema:
                type: object
                required: [libraries]
                properties:
                  libraries:
                    type: array
                    items: { $ref: "#/components/schemas/Library" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /series:
    get:
      operationId: listSeries
      summary: Lister les séries
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/LibraryId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page de séries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeriesPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /series/{seriesId}:
    get:
      operationId: getSeries
      summary: Détail d'une série
      tags: [catalog]
      parameters:
        - name: seriesId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Série et ses albums, dans l'ordre de lecture
          content:
            application/json:
              schema:
                type: object
                required: [series, comics]
                properties:
                  series: { $ref: "#/components/schemas/Series" }
                  comics:
                    type: array
                    items: { $ref: "#/components/schemas/Comic" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /comics:
    get:
      operationId: listComics
      summary: Lister les albums
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/LibraryId"
        - name: seriesId
          in: query
          schema: { type: string, format: uuid }
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/ComicState" }
        - name: readStatus
          in: query
          description: |
            Filtre sur la progression de lecture du compte courant. Omis ou
            vide, aucun filtre. `unread` inclut les albums jamais ouverts,
            qui n'ont aucune ligne de progression.
          schema:
            type: string
            enum: ["", unread, in_progress, read]
        - name: sort
          in: query
          description: |
            Ordre d'affichage. Défaut `recent`.

            Le curseur est lié au tri : le changer invalide la position
            précédente, et le serveur repart alors du début plutôt que de
            servir des résultats incohérents.
          schema:
            type: string
            enum: [recent, title, released]
        - name: folder
          in: query
          description: |
            Restreint à un dossier et à ses sous-dossiers. La chaîne vide
            désigne la racine de la bibliothèque, ce qui est distinct de
            l'absence du paramètre.
          schema: { type: string }
        - name: favorites
          in: query
          description: Ne retourne que les albums mis en favori.
          schema: { type: boolean }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page d'albums
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComicPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /accounts:
    get:
      operationId: listAccounts
      summary: Comptes du serveur
      tags: [accounts]
      responses:
        "200":
          description: Comptes
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items: { $ref: "#/components/schemas/Account" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createAccount
      summary: Ouvrir un compte
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                email: { type: string }
                password:
                  type: string
                  minLength: 12
                  description: |
                    Douze caractères minimum. La longueur est la seule règle :
                    exiger majuscules et chiffres pousse surtout à des variantes
                    prévisibles du même mot.
                role:
                  type: string
                  enum: [admin, user]
                  description: "Défaut : user."
                displayName: { type: string }
      responses:
        "201":
          description: Compte créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /accounts/{userId}/library-access:
    get:
      operationId: listAccountAccess
      summary: Bibliothèques explicitement ouvertes à un compte
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: Accès
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryGrants" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /libraries/{libraryId}/access:
    get:
      operationId: listLibraryAccess
      summary: Comptes autorisés sur une bibliothèque
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/LibraryIdPath"
      responses:
        "200":
          description: Accès
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryGrants" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: grantLibraryAccess
      summary: Ouvrir une bibliothèque à un compte
      description: |
        **Attention au modèle** : une bibliothèque SANS aucune autorisation
        explicite est visible de tous. Le premier accès accordé la referme donc
        pour tous les autres comptes.
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/LibraryIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
                canWrite:
                  type: boolean
                  description: Autorise le dépôt et la modification.
      responses:
        "200":
          description: Accès accordé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryGrant" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /libraries/{libraryId}/access/{userId}:
    delete:
      operationId: revokeLibraryAccess
      summary: Retirer l'accès d'un compte
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/LibraryIdPath"
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "204":
          description: Accès retiré
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Jeton d'accès obtenu via `POST /auth/login`. Durée de vie courte
        (15 min par défaut) ; à renouveler avec le refresh token.
...
  responses:
...
    Forbidden:
      description: Droits insuffisants
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationFailed:
      description: Validation échouée
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...